+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.8.0:
 - Added a new setting called "showLineNumbers", which states if the line numbers should be shown (defaults to true).
+- Renamed the public function "buildNewSyntaxElements" to "findAndBuildNewElements()".
 - 
 
 <br>